agent_beta/requirements.txt
flask
gunicorn
google-genai
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-gcp-trace